Just copy the CoC like I should have fuckin been doin already
CoC has two sorts--- P and T. P is the type of all types and P has type T.
Because of ∀-quantification over types in Rω, there is no compromise:
you must commit either to proper predicativity or proper impredicativity.

To illustrate, consider these translations.

  ⟦ τ₁ `→ τ₂ ⟧τ = Π ⟦ τ₁ ⟧τ (weaken ⟦ τ₂ ⟧τ)
  ⟦ `∀ κ τ ⟧τ = Π ⟦ κ ⟧κ ⟦ τ ⟧τ

The first quantifies over types, the second over kinds.  As expected in a
dependent language, these both translate to Π bound terms. This gets tricky when
typing Π in Ix; what should its sort be? I.e., what is σ in

  Δ ⊢ Π τ υ ⦂ σ

Trying to restrict ourselves to two levels---𝓤₀ and 𝓤₁---would imply
that the quantified type, τ, should equal the type of the Π-bound term, e.g.

  Δ ⊢ Π Nat Nat ⦂ 𝓤₀   because Nat ⦂ 𝓤₀

and

  Δ ⊢ Π 𝓤₀ Nat ⦂ 𝓤₁.  because 𝓤₀ ⦂ 𝓤₁.

But then what about the type of υ? If we are quantifiying over something in 𝓤₀,
it should be forced to stay in 𝓤₀. But recall the definition of rows.

  Σ (n : Nat). (Π (i : Ix i). 𝓤₀)

The body here has type 𝓤₁, but we have only quantified over things in 𝓤₁.
Okay, so actually then we need some lub business: the type of Π τ υ should
be (type of τ) ⊔ (type of υ). And okay, this is maybe not too bad if we limit
ourselves to *only* quantification over 𝓤₀ and things with type 𝓤₀. But...
we need to quantify over rows and row variables, which have type (at minimum)
𝓤₁. So you end up needing (if I'm correct) the full tower of levels.

On the other hand, we can copy CoC and just do this.
  (i)   Δ ⊢ 𝓟 ⦂ 𝓣
  (ii)  Δ ⊢ Π Nat Nat ⦂ 𝓟
  (iii) Δ ⊢ Π 𝓟 Nat ⦂ 𝓟