pfffffffffffffffffffffffffft
Scrap Mix5 and Pfft.agda. I tried to separate sorting and typing judgements
within a tagged representation. This was a bad idea. You may separate the two in the paper calculus
all you like, and should for clarity. But not here, where the tags of
overlapping pieces of syntax lead to unnecessary repeated syntax.