Refactor KEnv _,_ to _،_
It was long overdue to make KEnv _,_ an infixl operator,
but doing so causes Agda to fail to distinguish pairs from TVars.
This happens at parse-time, so type information is not available.
That is to say, even though I have written
   Type (Δ , κ) κ'
Agda sees
  (Δ , κ)
and asks: do I parse this to a pair or to a KEnv? And the parser
does not know that Type : KEnv ℓΔ → Kind ℓ → Set.

The only way I know of to fix this is to use a different operator.
So I have changed _,_ to _،_ for now.

(If in the future we want to change this symbol, it will be a simple
find-and-replace over ./Systems/Rome/.)