Restore stability to main
I need to switch to branch-based changes, as we're (we're? I'm) experimenting
in directions that ~~may fail~~ have failed. What's on main now:
- RecΣ, tie, ▿μ, et al are primitives.
- Meaning of these terms is defined in the semantic domain.

We (also I) suspect that ▿μ, tie, and RecΣ are completely admissable, and
therefore need not be defined in the semantic domain. I found this to be a
gargantuan pain in the ass so far. IMO, so long as you can show on pen and paper
that they are admissable, it doesn't matter too much if you retain them as
primitives for ease of definition. So, specifically, I would rather keep them
well-defined in the semantics for now and then try to derive them
syntactically... If that goes well, we can choose to remove them from the term
syntax.