Replace ⌈_⌉·_ and _·⌈_⌉ with ↑_ with _↑
The basic rationale here is that the lift operators really had nothing
to do with their arguments, so they should be stated in terms of the
function alone.

Along the way, I think this makes the kinding rules somewhat easier to
understand---there's no extra nonsense about the argument.

As confirmation, I didn't have to make changes in the index
calculus---the natural way to implement the previous operators really
only cared about the functions...