Remote System F projection
./Systems/F was the Rω development stripped of the row bits. If we need this
again, it will be simple enough to delete fields from a copy of Systems/Rome.

N.b. the founding intent of this repo (System-R-Etc) was to have a collection of
faux-extensible calculi leading from System F to Fω to Rω and so forth. Last I
checked, the extension problem is still a thing... Better to clean up trash for now.