This may be a dead end because of levels
I am afraid I need to equate levels, which cannot be done.

In the definition of the soundness relation, I must assert that two kinds are equal. These kinds have different levels, so I must likewise assert the two levels are equal. Maybe this is doable, but it's a massive pain in the ass.